replace sampe test by something working
