+++ ct-app/tests_endurance/test_sample_counter.py
@@ -1,0 +1,24 @@
+import logging
+
+from tools import getlogger
+
+from . import EnduranceTest, Metric
+
+log = getlogger()
+log.setLevel(logging.ERROR)
+
+
+class SampleCounter(EnduranceTest):
+    async def on_start(self):
+        self.counter = 0
+
+    async def task(self) -> bool:
+        self.counter += 1
+
+    async def on_end(self):
+        pass
+
+    def metrics(self):
+        self.metric_list = [
+            Metric("Counter", self.counter, cdt=f"== {self.duration * self.rate}")
+        ]
